Refactor event model
