+++ src/jupyterlabcontroller/models/domain/kubernetes.py
@@ -2,17 +2,18 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
+from copy import deepcopy
+from dataclasses import dataclass, field
 from enum import Enum
-from typing import Optional, Self
+from typing import Any, Self
 
 from kubernetes_asyncio.client import V1ContainerImage
 
 from .docker import DockerReference
 
 __all__ = [
+    "KubernetesEventData",
     "KubernetesNodeImage",
-    "KubernetesPodEvent",
     "KubernetesPodPhase",
 ]
 
@@ -85,26 +86,44 @@ class KubernetesPodPhase(str, Enum):
 
 
 @dataclass
-class KubernetesPodEvent:
-    """Represents an event seen while waiting for pod startup."""
+class KubernetesEventData:
+    """A helper class to capture the most useful data about a Kubernetes
+    Event and focus on a particular field within the event's involved object.
+    """
 
-    message: str
-    """Message in the Kubernetes event."""
+    type: str
+    raw_object: dict[str, Any]
+    name: str
+    kind: str
+    field: list[str] = field(default_factory=list)
+    missing_field: bool = True
+    value: Any = None
 
-    phase: KubernetesPodPhase
-    """Current phase of the pod."""
+    @classmethod
+    def from_kubernetes_event(
+        cls, event: dict[str, Any]
+    ) -> KubernetesEventData:
+        raw_object = event["raw_object"]  # This will exist (I think)
+        e_type = event["type"]
+        name = "<unknown name>"
+        kind = "<unknown kind>"
+        if "metadata" in raw_object and "name" in raw_object["metadata"]:
+            name = raw_object["metadata"]["name"]
+        if "kind" in raw_object:
+            kind = raw_object["kind"]
+        return cls(
+            type=e_type, name=name, kind=kind, field=[], raw_object=raw_object
+        )
 
-    error: Optional[str] = None
-    """Additional error accompanying this event (usually from the pod)."""
-
-    @property
-    def done(self) -> bool:
-        """`True` if the pod has started or definitively failed to start.
-
-        An unknown phase is considered a failure. The Kubernetes documentation
-        says that this can happen when the node on which the pod is supposed
-        to be running cannot be contacted, which is a sufficiently broken
-        state that we should consider the spawn a failure rather than waiting
-        to hope it will fix itself.
-        """
-        return self.phase != KubernetesPodPhase.PENDING
+    def reduce_by_field(self) -> None:
+        obj = self.raw_object
+        if self.field:
+            fldval = deepcopy(obj)
+            self.missing_field = False
+            for fld in self.field:
+                try:
+                    fldval = fldval[fld]
+                except (KeyError, TypeError):
+                    self.missing_field = True
+                    break
+            self.value = fldval
